removing proteins with low pLDDT score

+++ cif_download.py
@@ -1,4 +1,5 @@
 import requests
+from pathlib import Path
 
 proteins = []
 with open("PF04034_reduced.fasta", "r") as f:
@@ -12,6 +13,8 @@ with open("PF04034_reduced.fasta", "r") as f:
 
 for i in range(len(proteins)):
     protein = proteins[i]
-    URL = f"https://alphafold.ebi.ac.uk/files/AF-{protein}-F1-model_v3.cif"
-    tmp = requests.get(URL)
-    open(f"cif/AF-{protein}-F1-model_v3.cif", "wb").write(tmp.content)
+    my_file = Path(f"cif/AF-{protein}-F1-model_v3.cif")
+    if not my_file.is_file():
+        URL = f"https://alphafold.ebi.ac.uk/files/AF-{protein}-F1-model_v3.cif"
+        tmp = requests.get(URL)
+        open(f"cif/AF-{protein}-F1-model_v3.cif", "wb").write(tmp.content)
